% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexInput.R
\name{complexInput}
\alias{complexInput}
\title{A selectInput and textInput combined}
\usage{
complexInput(
  inputId,
  pickerOpts,
  textOpts,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
Try to encapsulate 2 Shiny inputs in a single HTML widget
which can save its state in the url, through bookmarking
}
\details{
TO DO: find a way to generalise this approach???
}
